requirements.txt
# torch installed separately in Dockerfile for CPU-only build
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
transformers>=4.35.0
supabase>=2.0.0
anthropic>=0.39.0
python-dotenv>=1.0.0
pydantic>=2.0.0
safetensors>=0.4.0
pdfplumber>=0.10.0
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0
seqeval>=1.2.2
datasets>=2.14.0
scikit-learn>=1.3.0
rapidfuzz>=3.5.0
